backend/pyproject.toml
[tool.poetry]
name = "team-principal-backend"
version = "0.1.0"
description = "TeamPrincipal — E-Sports Racing Management Platform (Backend API)"
authors = ["Bruno Strumendo <strumendo@gmail.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.0"
alembic = "^1.14.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.0.0"
python-multipart = "^0.0.18"
email-validator = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
